LockObjects & OrganizeArt
LockObjectsToWall.py and OrganizeArt.py are now pretty dang functional. I think LockObjectsToWall.py just needs to have the ui_styles.py properly applied to it (not sure why the buttons are not matching the give style).

OrganizeArt.py is the drag-and-drop wall that is missing from editorUI.py
I have been unable to get the features to work inside the provided window, but the UI functions well.