Add import/export from excel to artwork.py
Added an import and export method for excel sheets, this should eventually replace the old json method. I have not removed those functions as I didn't want to break anything. Theoretically the import and export from excel should allow for image functionality, but I haven't been able to test it. Also note the export only exports a single artwork as the method is self.export_to_excel. On the other hand, the import will get a list of artworks from an excel sheet. The import and export follow the __init__ format for the artwork class rather than the sample excel.