Update wall.py import/export
I have added an import and export for excel files to the wall class. I have left all of the old imports and exports for now so that I don't break other code. We definitely should remove the old methods later unless we still need them. Currently the export will take a wall and export its data, all of the artwork and artwork metadata to an excel sheet. The same is true for import, the functions will get all of the wall data, artworks and artwork metadata and output a wall object with all relevant data.