three decimal places fix
Not sure how you feel about the need for this. I did a find and replace for .2f and made them .3f (except prices obviously)

A quick check was not able to find anything wrong, but also I feel like this isn't needed before we turn this in. So maybe use your own best judgement before deciding if it should actually be pulled in.